Added First Tested Support for complete immutability
This is the first edition with support for making a builder completely
immutable. The copy architecture has been greatly simplified and all
setters have been inspected to ensure they do not mutate any contained
data members.

Due to a change in the interface of copy structure, the major version
has been incremented